Only navigate to source files if we have a path
According to the spec, "path" is an optional property of a Source, so
in some cases, stack frames may have a "source" but no "path"
inside (for example, if the stack frame is OS library code). We now
only try to open a source file if there is a path attached.