Require json; fix list type in json parser; prepend name; works now
json was used without requiring 'json. This was an oversight, fix this.

When using the patch, debugging did not work if multiple configurations were
used as the list of launch configurations acquired from the json property
"configurations" was an array. Fix this by binding json-array-type to 'list.

The last bug was that launch configuration names were not prepended to launch
configurations extracted from a launch.json file, resulting in a single launch
configuration called ':name' being shown. Fix this by prepending the launch
configuration's name (from its "name" json property) to it before returning the
result in the launch configuration provider.

The enhancement works now and can be used with dap-debug.