Add escaping; stricter variable regexes; fix env: bug
Variable blocks ${...} can now be escaped. \${...} always becomes ${...}.
Additionally, '}' can now be escaped in variable blocks. Previously, there was
no way to refer to a 'env:}' variable, for example. Now, this can be done with
${env:\}}.

All variable-related regular expressions now require that the entire variable
match. Previously, ${workspaceFolderFooBar} was equivalent to
${workspaceFolder}. Fix this, throwing an error in that case.

The ${env:...} variable had a bug causing a strange error to be thrown if the
environment variable references did not exist. Fix this, making it warn and
return the empty string instead.