Implement numbered prompting variables
To implement that feature, a two-pass architecture for variable expansion was
needed: first, gather all prompts. Then execute them, asking the user to enter
his values, and finally to the launch configuration everything with the prompted
values. Generalize dap-variables-expand-in-launch-configuration:
dap-variables-walk-launch-configuration. The new function recursively walks a
launch configuration, and expands variables within it with a function. This way,
the two-pass architecture could be implemented.

Add new hooks: dap-variables-pre-expand-hook, dap-variables-post-walk-hook and
dap-variables-post-expand-hook. They are called before expansion, after the
first walk-pass (that does not expand variables) and after the expansion
of a launch configuration, respectively. The post-expand hook is currently
unused. The first walk-pass also has a list of (REGEX . VALUE) mappings, just
like the final expansion pass. However, VALUE must always be functions and their
results are ignored.

Currently, only one variable is defined for this pass: ${NUMBER:PROMPT},
implemented in dap-variables--colon-prompt-var. It prompts the user with prompt
and stores the result in NUMBER. The value entered by the user can then be
accessed with ${NUMBER}, as often as needed. It gathers all variables of the
form described above into the list dap-variables-numbered-prompts. After the
first walking-pass, in the post-walk hook, implemented in function
dap-variables--do-prompts, this list is sorted by NUMBER (since prompt 1 has to
be asked before prompt 20, for example) and the user is asked to enter something
for every PROMPT. On the left, a progress statistic is shown, telling the user
how many questions are still left. Duplicate PROMPT-NUMBER pairs are handled
with a warning, one of them (it is undefined which, due to --sort not being
stable) winning. History is remembered, however only shared between all
variables. An attempt of implementing per PROMPT history is still in the code,
however commented.