Use unwind-protect to reset dap-variables-numered-prompts
Previously, two lambdas, in pre-expand-hook and post-expand-hook, were used to
reset dap-variables-numered-prompts, to not waste conses and to not interfere
with the next prompt run. Replace this is a new function,
dap-variables--do-prompts-reset, which calls dap-variables--do-prompts in an
unwind protect form that always resets dap-variables-numered-prompts to nil.
This way, even if the user cancels the debug operation at any point in time,
dap-variables-numbered-prompts will be nil, not interfering with prompt runs and
not wasting conses.

Additionally, fix some indentations (M-x mark-whole-buffer M-x indent-region).