Add escaping; stricter variable regexes; fix env: bug (#333)
* Add escaping; stricter variable regexes; fix env: bug

Variable blocks ${...} can now be escaped. \${...} always becomes ${...}.
Additionally, '}' can now be escaped in variable blocks. Previously, there was
no way to refer to a 'env:}' variable, for example. Now, this can be done with
${env:\}}.

All variable-related regular expressions now require that the entire variable
match. Previously, ${workspaceFolderFooBar} was equivalent to
${workspaceFolder}. Fix this, throwing an error in that case.

The ${env:...} variable had a bug causing a strange error to be thrown if the
environment variable references did not exist. Fix this, making it warn and
return the empty string instead.

* Implement numbered prompting variables

To implement that feature, a two-pass architecture for variable expansion was
needed: first, gather all prompts. Then execute them, asking the user to enter
his values, and finally to the launch configuration everything with the prompted
values. Generalize dap-variables-expand-in-launch-configuration:
dap-variables-walk-launch-configuration. The new function recursively walks a
launch configuration, and expands variables within it with a function. This way,
the two-pass architecture could be implemented.

Add new hooks: dap-variables-pre-expand-hook, dap-variables-post-walk-hook and
dap-variables-post-expand-hook. They are called before expansion, after the
first walk-pass (that does not expand variables) and after the expansion
of a launch configuration, respectively. The post-expand hook is currently
unused. The first walk-pass also has a list of (REGEX . VALUE) mappings, just
like the final expansion pass. However, VALUE must always be functions and their
results are ignored.

Currently, only one variable is defined for this pass: ${NUMBER:PROMPT},
implemented in dap-variables--colon-prompt-var. It prompts the user with prompt
and stores the result in NUMBER. The value entered by the user can then be
accessed with ${NUMBER}, as often as needed. It gathers all variables of the
form described above into the list dap-variables-numbered-prompts. After the
first walking-pass, in the post-walk hook, implemented in function
dap-variables--do-prompts, this list is sorted by NUMBER (since prompt 1 has to
be asked before prompt 20, for example) and the user is asked to enter something
for every PROMPT. On the left, a progress statistic is shown, telling the user
how many questions are still left. Duplicate PROMPT-NUMBER pairs are handled
with a warning, one of them (it is undefined which, due to --sort not being
stable) winning. History is remembered, however only shared between all
variables. An attempt of implementing per PROMPT history is still in the code,
however commented.

* Implement per prompting variable histories

The history of each prompting variable (${N:PROMPT}) is now distinct. This was
implemented with a clever hack, see dap-variables--temp-hist.

* Address PR comment: remove progn

* Integrate with savehist

Save per-prompting-variable histories with savehist. To do so, add the variable
that stores them to savehist-additional-variables. If that variable does not
exist, do nothing, handling it gracefully.

* Make launch.json the default configuration provider

Previously, the launch configurations in dap-debug-template-configurations were
shown before that of those in the project's launch.json file. Change that,
showing those in the project's launch.json file first.

* Set dap-variables-numbered-prompts to nil in post-walk-hook

If it is set at initialization, its value will linger around until the next call
to dap-debug. This is a waste of memory, as it is never used again. Set it '() in
post-walk-hook, after dap-variables--do-numered-prompts, so that its memory can
be immediately reclaimed afterwards.

* Also set dap-variables-numbered-prompts to nil before walking

If the user cancels dap-debug before it starts, dap-variables-numered-prompts
would be non-nil, causing leftover prompts to be queried. Set it to nil before
debugging, also.

* Use unwind-protect to reset dap-variables-numered-prompts

Previously, two lambdas, in pre-expand-hook and post-expand-hook, were used to
reset dap-variables-numered-prompts, to not waste conses and to not interfere
with the next prompt run. Replace this is a new function,
dap-variables--do-prompts-reset, which calls dap-variables--do-prompts in an
unwind protect form that always resets dap-variables-numered-prompts to nil.
This way, even if the user cancels the debug operation at any point in time,
dap-variables-numbered-prompts will be nil, not interfering with prompt runs and
not wasting conses.

Additionally, fix some indentations (M-x mark-whole-buffer M-x indent-region).

* Make unused _ parameter in dap-variables--do-prompts-reset optional

The function could be useful in contexts other than post-walk-hook, so make _
optional, since it isn't used anyway.