Handle comments like /***/ correctly in dap-launch-remove-comments
Comments that ended with and even number of stars and a slash were not removed.
This was because of the naive regex that matched * followed by anything that is
not a slash. This would cause the following star to be matched, even though it
should end the comment. A solution would have been to use a look-ahead assertion
like ?!/. However, Emacs' regex flavour does not support them, and POSIX doesn't
either. I solved the problem using backtracking: now the body of a C89-style
comment can either contain any character that is not a * or a sequence of one or
more stars followed by something that is not a / (this part will backtrack if a
slash follows), ending in one or more stars followed by a slash.

Add a test that checks if comments like /***/ are removed correctly.