Handle comments like /***/ correctly (#342)
* Support launch.json files in a .vscode subdirectory

Extend dap-launch-find-launch-json to try looking for the launch.json file in
the project-root's .vscode subdirectory if there is no launch.json in the
project root.

Remove (require cl-lib), as it was not used.
Use and require the f library to build launch.json paths (f-join).

* Prefer the launch.json file in the .vscode subfolder

* Support comments in launch.json files

Add a new function, dap-launch-delete-comments, that deletes all C-style
comments in the current buffer, respecting string constants. Use it in
dap-launch-find-launch-json, which now no longer calls json-read-file but
inserts the launch.json file into a temporary buffer, calls
dap-launch-delete-comments and then parses the buffer.

* Add tests for dap-launch's new comment deletion function

Test dap-launch-delete-comments in tests/dap-launch-tests.el. The tests cover
expansion in strings, nested comments, normal comments and strings after
comments.

* Handle comments like /***/ correctly in dap-launch-remove-comments

Comments that ended with and even number of stars and a slash were not removed.
This was because of the naive regex that matched * followed by anything that is
not a slash. This would cause the following star to be matched, even though it
should end the comment. A solution would have been to use a look-ahead assertion
like ?!/. However, Emacs' regex flavour does not support them, and POSIX doesn't
either. I solved the problem using backtracking: now the body of a C89-style
comment can either contain any character that is not a * or a sequence of one or
more stars followed by something that is not a / (this part will backtrack if a
slash follows), ending in one or more stars followed by a slash.

Add a test that checks if comments like /***/ are removed correctly.

* Make dap-launch.el file header comment use ;;; instead of ;;