Feature/dap debug restart remove old (#345)
* Extend dap-debug-restart to allow deleting the old session

Extend dap-debug-restart to take an optional argument (in interactive mode it is
the prefix argument) that, if set, causes the old session to be deleted.

* Use dap-delete-session conditionally instead

* Make dap-debug-restart's default keeping behaviour configurable

The user can now configure whether dap-debug-restart should keep the old session
around with the variable dap-debug-restart-keep-session (the default value being
the old behavior). The prefix argument now overrides it, instead of settings
the behavior.