Improve variable expansion; add tests for it (fixes #346) (#351)
* Improve variable expansion; add tests for it

Handle lists generically in `dap-variables-walk-launch-configuration':
previously, alists were detected and handled with -all? #'consp, which is not
ideal, as it precludes lists of lists from being expanded correctly; however,
worse than that, all other types of lists were automatically assumed to be
plists. This led to a bug if a list with an odd number of elements was to be
expanded: it would end in nil. Fix this by considering any list a normal list,
and by handling conses whose `cdr' is not a list specially instead.

Also support vectors correctly now (yield a vector if a vector is given).

Test the variable expansion code in tests/dap-variables-test.el, with a set of
tests exercising previous issues.

* Ensure with tests that functions are unmodified during expansion

Also clarify the DOCSTRING parameter of `dap-variables-test--compare' in its
docstring.

* Fix dap-variables-test-lambda-untouched

Two different lambdas, even if they are equivalent in their string
representation, are not `equal'. Use #'identity in
`dap-variables-test-lambda-untouched' instead and rename it to
`dap-variables-test-function-untouched'.