bugfix: expand variables before passing them to debug providers
Previously, variables were expanded in `dap-start-debugging`, which was called
after passing the configuration to a debug provider in `dap-debug`. This caused
issues with the python debug adapter, which passes the :args field as a quoted
shell argument, bypassing variable expansion. Add a new function
`dap-start-debugging-noexpand`, which is like `dap-start-debugging` but does not
expand variables.