dap-variables: support ${command:pickProcess}
Implement ${command:pickProcess} in terms of a new function,
`dap-variables--pick-process`, prompting the user to select a process in a
completing-read minibuffer, returning the PID as a number.

Extend `dap-variables--eval-poly-type` to support numbers via two-stage
evaluation and use pcase instead of cond in it. This way,
`dap-variables--buffer-current-line` can be replaced with just
line-number-at-pos.

No longer require 'dash, since it isn't used anywhere.