Implement "runInTerminal" for dap-mode
dap-mode supports both types of "runInTerminal" terminals: "external" and
"integrated", the latter implemented using vterm. Split `dap--make-response'
into `dap--make-success-response' and `dap--make-error-respones', all of which
now also require the command that was executed as a string (as per the spec).
`dap--make-response' is now a more "raw" function, taking the success flag also.