Implement :console "integratedTerminal"/"externalTerminal" (#362)
* Rebase against master

* WIP - external

* WIP - support vterm (untested)

* Implement "runInTerminal" for dap-mode

dap-mode supports both types of "runInTerminal" terminals: "external" and
"integrated", the latter implemented using vterm. Split `dap--make-response'
into `dap--make-success-response' and `dap--make-error-respones', all of which
now also require the command that was executed as a string (as per the spec).
`dap--make-response' is now a more "raw" function, taking the success flag also.

* Make internal terminal configurable

Add a new defcustom, `dap-internal-terminal', which can be used to set the
function that is to run and display the integrated terminal. Implementations are
available for vterm, async-shell and to auto-detect between the two (preferring
vterm).

* dap-external-terminal: use -T instead of -display to set xterm title

In dap-mode.el: dap-external-terminal defaults to using xterm, which specifies
the title with -T. Using -display to set the title is wrong, so use -T instead.

* dap-mode: external: address review

Address some of the minor (easy to fix) issues in the PR review:

- dap-internal-terminal-vterm: don't pass one to (vterm-mode). It is a derived
  mode, so passing an argument to it yields "wrong number of arguments" errors.
- dap-internal-terminal-auto: check (fboundp) for 'vterm, not 'vterm-mode. The
  latter isn't autoloaded, so vterm would only be selected if it is loaded (it
  should be selected if it is just installed)

* dap-mode: handle failure to start external terminal gracefully

When the user's external terminal (`dap-external-terminal') is requested and
dap-mode fails to start it for whatever reason (most of the time because its not
installed), make a message using `with-demoted-errors' and use the built-in one
instead.

* dap-mode: improve warning on failure to launch external terminal

Now also tell the user some steps they can take to fix the issue (install xterm,
configure `dap-external-terminal').

* dap-mode.el: fix some flycheck warnings

Defvar global variables that are overriden using let.

Co-authored-by: Ivan Yonchovski <yyoncho@gmail.com>