dap: external terminal: isolate with-demoted-errors to start-process
In the block responsible for launching the external terminal, move
dap--send-message outside of the with-demoted-errors block. This way, no
misleading error message will be presented to the user if dap--send-message
fails.

Minor bugfix: in the with-demoted-errors format, use %S instead of %s. The
latter is only for strings, which is wrong (errors can be sexps).

Replace a let* with a let, since there is only one binding.