dap-python: debugpy: support built-in templates
The built-in python templates set many keys to nil, which causes initialization
errors with debugpy (it expects them to be unspecified instead). When debugpy is
used, check such keys for nil, and if so, remove them from the configuration.
This way, the built-in templates work with both debugpy and ptvsd.