Handle per-platform launch attributes
In debug configurations, VSCode allows setting some properties only for certain
platforms. In JSON objects, "<platform-name>": {attrs...} overrides attrs in the
containing dictionary.

Support this in dap-variables, and make use of the new support in
`dap-variables-expand-in-launch-configuration'. A side effect of this change is
that the latter function now only supports property lists.