dap-mouse: support triggering manually
Refactor the value rendering of `dap-mouse' into a new function:
`dap-ui-render-value'.

The new `dap-tooltip-at-point' function creates a dap-mouse posframe under the
cursor, and can be used without `dap-tooltip-mode'. Extend `dap--resp-handler'
to take an error callback. which is leveraged in `dap-tooltip-at-point' to make
the error handling more explicit.

Add functions to evaluate expressions in a new `treemacs' buffer:
`dap-ui-eval-in-buffer', `dap-ui-eval-variable-in-buffer', based on the new
`dap-ui-render-value' function.