Add ui-display-frame to dap--debug-session struct
Prior to this commit, when the debugger stops (e.g. breakpoint), it
spawns the ui elements in whatever is `(current-frame)`.  This changes
so that the UI display frame is set when the user starts or switches
to a session.

In particular the prior functionality is a bug for `exwm` users
debugging a program that produces a floating graphical window, which
will be contained in an Emacs frame under `exwm`.  Often a breakpoint
would be triggered by user interaction so that frame has focus.