`dap-go': bump vscode-go version
Version 0.15.0 of vscode-go does not handle maps correctly, adding nulls to
them. This bug is fixed in version 0.22.1, so bump it in `dap-go-setup'.

The layout of the extension changed: the executable that needs to be run is no
longer goDebug.js, bug dist/debugAdapter.js. As such, users will need to
manually update `dap-go' using `C-u dap-go-setup'.

Fixes #399.