dap-go: fix type error when splitting :args field (#460)
If the :args field is already present in a given configuration, the
code tries to call `split-string` on that field, which is appropriate
to do once but will fail on subsequent calls, such as those
encountered when running `dap-debug-last`. Changing the conditional
here from a nil check to a `stringp` check worked for me.