dap-ui: Fixed issue where control icon background was unspecified
This resolves #462.

If a user starts `dap-debug`, dap-mode creates a set of icons used to
do common debug actions (step, continue, etc). These icons are created
using `propertize` to make the text that displays the images on a
background.

The existing setting uses the background color from the `fringe` face
to set the background of the icon. The return value for the `face-attribute`
function is not guaranteed to be specified. When the `:background` is
unspecified, the space is made for the icons but nothing is displayed.

This PR defines a face to control the background color used for the
debug icons. It defaults to using the background color defined in the
`fringe` face, but will inherit from the `default` face if the
`:background` attribute is unspecified in `fringe`.