dap-gcc-lldb: fix setup (get from GitHub releases) (#618)
This problem has been consistently present since last year, as VSCode
Marketplace blocks direct downloads, It was impossible to install a
VSCode extension with the available helper functions.

This commit fixes this issue for the webfreak.debug extension, as the
extension's GitHub repo delivers releases for the extension. I used this
link to download it (inspired by `dap-cpptools-setup`).