Protect breakpoint persistence from print-* vars
The output of printing functions (including 'prin1' and, by
extension, 'format' with a '%s' sequence) is affected by a set of
output variables that both end-users and libraries can tweak.

This is fine in the more common use of these functions to format
messages, but the problem is that the same functions are also the
most convenient way to serialize/persist data.

Serialization should be robust to non-default values of
'print-length' and 'print-level' in particular, as setting these to
small numbers can lead to accidental data loss (and in some cases
unreadable output).

Starting in Emacs 29, 'prin1' and co. gained a new optional argument
precisely to protect against this problem.  In the meantime, this
patch ensures that 'print-length' and 'print-level' are set to their
default values during serialization.