fix: fixes error when a utility frame is the current frame
On Mac OS X, small frames that don't seem to have any windows may be
selected when the process filter is processed.  By walking up the
parent-frame list, we can get to the frame that the user is actually
able to use, and thus the call to (get-mru-window) does not return nil
and the whole process filter aborts.

Should fix bug #828